refactor: temporarily remove flight generator for model changes
